### 2024-10-22 #7394 @editorialbot
[PRE REVIEW]: PointClouds.jl: Fast & flexible processing of lidar data
**Submitting author:** @mfsch (<a href="http://orcid.org/0000-0002-7880-9913">Manuel F. Schmid</a>)
**Repository:** https://github.com/efpl-columbia/PointClouds.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.4.0
