### 2024-11-04 #7394 @mfsch
@kthyng: Thanks for your message – we are preparing an abbreviated version of the paper, which we should be able to submit soon along with a list of suggested reviewers.
