### 2024-11-05 ceafff7
Add note to affiliations
