### 2024-12-10 7a2cbad
Revise JOSS draft to reduce word count
