+++ joss/paper.md
@@ -25,32 +25,21 @@ bibliography: references.bib
 
 [^1]: Marco G. Giometto holds concurrent appointments as Assistant Professor at Columbia University as Amazon Visiting Academic. This paper describes work performed at Columbia University and is not associated with Amazon.
 
+
 # Summary
 
 PointClouds.jl is a toolbox for working with point-cloud data that lets users implement complex point-processing workflows.
-It includes functionality for working with the LAS file format (and its compressed LAZ variant) that is commonly used to save and distribute geospatial point-cloud data obtained from lidar scans.
-Such scans are obtained with a laser instrument, often mounted on an airplane, that shoots out light pulses in quick succession and measures the return time of those pulses to reconstruct the points at which the signal was reflected.
-This results in a “point cloud” of 3D coordinates and various additional attributes associated with each point, such as the intensity of the reflected signal.
-Extracting useful information from that data generally requires a series of processing steps that extend, filter, or transform the point cloud.
-PointClouds.jl aims to provide a comprehensive set of functionality to build such processing pipelines.
-As lidar scans are becoming more commonplace, large datasets are being made publicly available online.
-The lidar point-cloud data collected by the United States Geological Survey as part of their 3D Elevation Program (3DEP) now provides near-complete coverage of the United States at a sub-meter scale, for example.
-PointClouds.jl also includes functionality to query such datasets and download the data for any region of interest, allowing a spatial analysis pipeline to be applied to any location within the current coverage.
-These capabilities make PointClouds.jl a fast and flexible tool for finding, loading, and analyzing lidar data.
+Geospatial point-cloud data is usually obtained with a “lidar” instrument that scans the surface with high-frequency laser pulses, and large datasets covering entire countries are being made publicly available online.
+PointClouds.jl includes functionality to query such datasets, load and save data in the commonly used LAS/LAZ file format, and extract useful information through a series of processing steps that extend, filter, or transform the point-cloud data.
 
 
 # Statement of need
 
-Remote sensing allows us to obtain a wealth of information about the physical environment.
-While many remote sensing technologies produce raster data (e.g. satellite imagery), methods such as lidar (light detection and ranging) and photogrammetry instead identify a collection of points on the land surface, producing a “point cloud” with the 3D coordinates and additional point attributes.
-Sensor platforms ranging from satellites to mobile phones generate point-cloud data that can span entire countries or more targeted high-fidelity scans.
-Thanks to falling acquisition costs and newly emerging platforms such as unmanned aerial vehicles and self-driving cars, point-cloud data is more readily available than ever before.
-
-It hence comes as no surprise that point-cloud data has found its way into numerous applications in industry, government, and academia.
+Point-cloud data has found its way into numerous applications in industry, government, and academia.
 Examples include computer vision for robotics and autonomous vehicles, generation of digital terrain & surface models, forestry and agriculture, archaeological surveys, geomorphology, and inspection/surveillance of infrastructure.
-Whenever an application requires digital data on the geometry of real-world surfaces, acquiring point-cloud data may be an attractive option.
+Thanks to falling acquisition costs and newly emerging sensor platforms, point-cloud data is more readily available than ever before.
 
-While some of these goals may be achieved by visualizing the point-cloud data and inspecting it manually, most applications rely on a series of automated processing steps to derive actionable information from the raw point-cloud data.
+Most applications rely on a series of automated processing steps to derive actionable information from the raw point-cloud data.
 These steps may include noise removal, subsetting, classification, object detection, rasterization, triangulation, and conversion between various data formats.
 Some processing steps may also make use of additional data obtained from complementary sensing technologies such as multispectral cameras or rely on pre-existing databases of geospatial information.
 
[…]
