### 2024-12-10 be3971b
Revise JOSS draft to reduce word count
