+++ joss/references.bib
@@ -52,3 +52,11 @@
   howpublished = {Presented at the AGU Annual Meeting, 11 – 15 Dec, San Francisco, CA},
   publisher = {American Geophysical Union}
 }
+
+@misc{Schmid:2024,
+  author = {Schmid, Manuel F. and Giometto, Marco G.},
+  title = {Towards an integrated model for operational microscale simulations},
+  year = {2024},
+  howpublished = {Presented at the AGU Annual Meeting, 9 – 13 Dec, Washington, D.C.},
+  publisher = {American Geophysical Union}
+}
