### 2024-12-10 #7394 @mfsch
@kthyng: I’ve pushed a shorter version of the manuscript to the `joss` branch of the PointClouds.jl repository.

Possible reviewers: evetion, visr, simondanisch, rafaqz, wkearn, lheagy, ruillercoaster
