### 2025-03-26 #7394 @kthyng
:wave: @amadeusine, @atzberg, @kellertuer, @zafeirakopoulos Are you interested in reviewing this submission for the Journal of Open Source Software (JOSS)? We ask for reviews to start in the next 2 weeks and aim to finish within 4-6 weeks. Our reviews are checklist-driven, please check our docs for more information: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
