### 2025-03-26 #7394 @kthyng
@kellertuer Thank you for your response. I try to avoid asking people to review when they are already doing so, but I don't see you listed (unless you mean a review with a different journal?). If you are currently doing another JOSS review could you send me the link so we can check why our database isn't finding it?
