### 2025-04-01 0c034e2
Update license and attribution text
