### 2025-04-04 efpl-columbia/PointClouds.jl#4 @mfsch
ScienceBase queries appear broken – can we use a different service?
Over the last months, ScienceBase has been unavailable from time to time, causing the `gettiles` functionality to stop working. We already have a warning about that [in our documentation](https://docs.mfsch.dev/PointClouds.jl/data-sources#USGS-3D-Elevation-Program-(3DEP)).

However, now it seems that ScienceBase is more permanently broken: The query generated in the [tutorial](https://docs.mfsch.dev/PointClouds.jl/tutorial) produces a [400 error](https://www.sciencebase.gov/catalog/items?format=json&fields=id%2Ctitle%2CwebLinks%2Cdates%2Cspatial&max=100&offset=0&filter=parentId%3D4f70ab64e4b058caae3f8def&order=asc&sort=dateCreated&filter=spatialQuery%3DPOINT%28-77.036389%2038.897778%29) and even just [listing all lidar entries](https://www.sciencebase.gov/catalog/items/get?parentId=4f70ab64e4b058caae3f8def) errors.

As a result, `gettiles` is pretty much broken. While the [file server](https://rockyweb.usgs.gov/vdelivery/Datasets/Staged/Elevation/LPC/) seems to work fine, we need a different way to perform the spatial query to obtain download links for LAZ tiles for the target coordinates.

Some options:
- Find a different service that can perform these queries for us.
- Build a local database (full or partial) of available lidar tiles by either directly reading the extent data from LAZ headers or finding and parsing metadata documents.
- Build such a database once/sporadically and make it available online, then download/query that database from within the package.
- Leave it broken and hope that USGS will fix it at some point.

Eventually it would be good to have some functionality in the package to build a local catalog of LAS tiles from a list of files or URLs, but building such a list for the whole dataset of ~7M tiles is not ideal. Using a different method to find the 3DEP “work unit” for an area and then cataloging the files just for that unit may be acceptable, but ideally we could avoid doing that work.
