### 2025-04-04 efpl-columbia/PointClouds.jl#4 @mfsch
Some potential data sources:
- [TNM Access](https://apps.nationalmap.gov/tnmaccess/) calls out to the ScienceBase API, so it has the same issues.
- On the file server, there is a ~2.4GB `WESM.gpkg` file that contains polygons and web links for the work units, but not for individual tiles.
- Work units should also have a tile index somewhere on the file server ([example](https://rockyweb.usgs.gov/vdelivery/Datasets/Staged/Elevation/metadata/NY_New_York_CMGP_SANDY_LiDAR_15/NY_CMPG_2013/shapefiles/tile_index/)), but it does not appear that this is at a predictable location.
- There is an ArcGIS server [here](https://index.nationalmap.gov/arcgis/rest/services/3DEPElevationIndex/MapServer/8) that also has the work units and allows spatial queries.
- The [Lidar Explorer](https://apps.nationalmap.gov/lidar-explorer/#/) service allows browsing the data and shows the work unit for a given area (probably using the same data as in the WESM.gpkg file), but the part that is supposed to show lidar tiles for an area is also broken since it is also calling out to TNM Access.
