### 2025-04-04 #7952 @mfsch
Just a heads-up for the reviewers: It appears that the online tile query feature is currently not working, because the USGS ScienceBase service we rely on is having problems. I’ve opened [an issue](https://github.com/efpl-columbia/PointClouds.jl/issues/4) for this.
