### 2025-04-15 efpl-columbia/PointClouds.jl#4 @mfsch
ScienceBase queries appear to be working again. However, now there is an error in the SSL configuration of the Rockyweb server where the LAZ files are hosted. I’ve reached out to USGS about this, and they informed me that they recently changed the architecture of that server and will forward the issue to the appropriate team.

In the meantime, defining the following method in your code should get downloads working again:

```
function PointClouds.DataSources.fetchtile(tile::PointClouds.DataSources.PointCloudTile{PointClouds.DataSources.ScienceBase})
  dir = PointClouds.DataSources.tiledir(tile)
  path = joinpath(dir, PointClouds.DataSources.filename(tile))
  if !isfile(path)
    @info "Tile `$path` not cached, downloading…"
    mkpath(dir)
    PointClouds.DataSources.HTTP.download(PointClouds.DataSources.uri(tile), path; require_ssl_verification = false)
  end
  path
end
```
