### 2025-04-15 #7952 @atzberg
Overall the package looks good.  The documentation is organized and clear.    The installation mostly works.  However, it did require a few manual adjustments to get the tests to work.  We report below a few errors we encountered.   We also provide a few suggestions that might help users avoid similar issues.

1. When following README.md, I had the following error. Perhaps I needed to first copy data to a specific location, then call this function to load.  I was expecting this to work directly from an examples folder or similar that contained some data.  I imagine researchers in this field can find such data and formats.  Perhaps add a note, or link to test data, so example loads more easily.

Error: “
julia> pts = LAS("USGS_LPC_Sandy_Supplemental_NCR_VA_MD_DC_QL2_LiDAR_18SUJ322306.laz")
ERROR: Could not load LAS data from `USGS_LPC_Sandy_Supplemental_NCR_VA_MD_DC_QL2_LiDAR_18SUJ322306.laz`
Stacktrace:
 [1] error(s::String)
   @ Base ./error.jl:35
 [2] LAS(input::String; kws::@Kwargs{})
   @ PointClouds.IO ~/.julia/packages/PointClouds/UhJj4/src/las-main.jl:505
 [3] LAS(input::String)
   @ PointClouds.IO ~/.julia/packages/PointClouds/UhJj4/src/las-main.jl:495
 [4] top-level scope
   @ REPL[8]:1
“

2. When running the tests, there was also an error that occurred (below).  I was able to manually correct this by installing an additional package, but may want to add this to the list of dependencies.

"ERROR: LoadError: ArgumentError: Package AllocCheck not found in current path.
- Run `import Pkg; Pkg.add("AllocCheck")` to install the AllocCheck package."

Once I manually installed this package,  then running tests worked.  They returned

Test Summary:  | Pass  Total     Time
PointClouds.jl | 2148   2148  1m08.5s

Suggestions:

- Add an "examples" folder with some data files and easy to run scripts that demonstrate functionality of the package.  This would help facilitate use of the package.

- Add some notes on common errors or issues like the above and remedies, if these are things that are unable to be resolved automatically by adjusting dependency list or similar solutions.

Overall, this looks like a useful package for processing lidar point-cloud data files in Julia.
