### 2025-04-16 57f1a93
Update AllocCheck version for Julia 1.11 compatibility
