+++ Project.toml
@@ -16,7 +16,7 @@ Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
 Proj = "c94c279d-25a6-4763-9509-64d165bea63e"
 
 [compat]
-AllocCheck = "0.1.3"
+AllocCheck = "0.2.2"
 BaseDirs = "1.2"
 DataFrames = "1.6"
 Dates = "1.10"
