### 2025-04-16 0baee09
Improve wording of quickstart in README
