### 2025-04-16 #7952 @mfsch
@atzberg: Thank you for taking a look at PointClouds.jl and for your feedback! To address the two issues you’ve encountered:

1. This quickstart example is meant to show how to get going with an existing LAS file. I’ve adapted the wording to make this more clear. Did you see the tutorial that is linked just below? This shows a more complete workflow that includes downloading the data first. Note that because of the [recent issue](https://github.com/efpl-columbia/PointClouds.jl/issues/4) with the USGS servers, the `gettiles` command will fail unless you apply the workaround that is mentioned in the issue.
2. The test dependencies are listed in the [`[extras]` section](https://github.com/efpl-columbia/PointClouds.jl/blob/822d88a08b4747dce6e4eadf9b383b1f2bc83bea/Project.toml#L33) of the `Project.toml` file so they are only installed when someone is running the tests rather than for every user of the package, as explained in the [Julia docs](https://pkgdocs.julialang.org/dev/creating-packages/#Test-specific-dependencies). You should be able to run the tests without manually installing AllocCheck by [running the tests through Pkg](https://pkgdocs.julialang.org/dev/managing-packages/#Testing-packages), e.g. by running `pkg> test PointClouds` (after `pkg> add PointClouds`) in the [Pkg REPL](https://pkgdocs.julialang.org/v1/getting-started/#Basic-Usage). Although testing this locally I stumbled into [this issue](https://github.com/JuliaLang/AllocCheck.jl/issues/80). I’ve fixed this by updating AllocCheck to the latest version and published a new version, so it should work now even when running Julia 1.11.

As for the examples folder, I would prefer to have usage examples in the documentation so they can be part of the automated tests using [doctests](https://documenter.juliadocs.org/stable/man/doctests/#Doctests). If you feel that the current tutorial does not cover enough of the functionality, I would rather add more examples there. Are there specific aspects of the usage that you would like to have better examples for?
