### 2025-04-16 #7952 @atzberg
@mfsch:  Thanks for this additional information.   I now see the link to the more in-depth tutorial within the docs.  That is fine to keep it with the docs to allow for auto-testing.  I think the coverage of functionality looks good, so no need for additional examples.  The in-depth tutorial is detailed and clear.

In summary, the package looks good and provides useful tools for processing lidar point cloud data in julia.

I've now completed the review here and it looks good to publish.
