### 2025-04-26 efpl-columbia/PointClouds.jl#4 @evetion
> In the meantime, defining the following method in your code should get downloads working again:
>
> ```
> function PointClouds.DataSources.fetchtile(tile::PointClouds.DataSources.PointCloudTile{PointClouds.DataSources.ScienceBase})
>   dir = PointClouds.DataSources.tiledir(tile)
>   path = joinpath(dir, PointClouds.DataSources.filename(tile))
>   if !isfile(path)
>     @info "Tile `$path` not cached, downloading…"
>     mkpath(dir)
>     PointClouds.DataSources.HTTP.download(PointClouds.DataSources.uri(tile), path; require_ssl_verification = false)
>   end
>   path
> end
> ```

While it's bad practice to disable ssl verification, maybe the above could be merged in the codebase and be released? Now that code part + tutorial will just not work at all.
