### 2025-04-26 #7952 @evetion
This is a lovely package, great to see such leaps in the ecosystem. I've read the paper, documentation and ran some code examples and tests without issue.

I didn't yet check all the boxes though:

- The documentation still needs a statement of need (could be copied from the paper).
- Community/contribution guidelines are missing
- State of the field is discussed outside of Julia, but the Julia parts are glanced over. It might be good to mention there's isolated functionality in LasIO.jl, LazIO.jl and PointCloudRasterizers.jl that this package effectively replaces. Furthermore, there's the concurrent development of LasDatasets.jl. Full disclosure, I'm the author of LazIO and PointCloudRasterizers.
- References: I think it would be good to also cite the LAS standard, and the laszip format (paper by the late Martin Isenburg).

I will create some further issues on the repo with some feature requests and suggestions, unrelated to this review.
