### 2025-04-27 efpl-columbia/PointClouds.jl#5 @evetion
Consider supporting COPC
Hi there, this is triggered by (but not needed for) the review at https://github.com/openjournals/joss-reviews/issues/7952.

It would be lovely to support [COPC](https://copc.io/): cloud optimized point cloud. It's LAZ, but with some specific ordering/chunking and a built-in index. It would be easier to stream laz datasets then, such as those at https://planetarycomputer.microsoft.com/dataset/3dep-lidar-copc.

That said, ideally we would use a new library like https://github.com/RockRobotic/copc-lib and https://github.com/hobuinc/laz-perf to do so, replacing the laszip_jll. It would require making a C library on top of those C++ libraries and compiling it, or doing some CxxWrap.jl coding.
