### 2025-04-27 efpl-columbia/PointClouds.jl#6 @evetion
Integrating with the JuliaGeo ecosystem
Hi there, this is triggered by (but not needed for) the review at https://github.com/openjournals/joss-reviews/issues/7952.

It would be great if we can further integrate PointClouds.jl the rest of the Julia ecosystem, using https://github.com/JuliaGeo/GeoInterface.jl/ (so it's easy to convert to/from other Point geometries) and https://github.com/JuliaGeo/GeoFormatTypes.jl. (so an EPSG code is not an Integer, but its own struct).

A second point of attention could be using Rasters.jl, maybe via a package extension, so it becomes very easy to save a RasterizedPointCloud as a geotiff or similar.
