+++ src/DataSources.jl
@@ -6,6 +6,7 @@ using Dates: Dates
 using BaseDirs: BaseDirs
 using HTTP: HTTP
 using JSON3: JSON3
+using Proj: Proj
 using ..IO: IO
 import ..AbstractPointCloud
 
@@ -20,7 +21,7 @@ end
 PointCloudTile(src::Type{S}, data::D) where {S,D} = PointCloudTile{S,D}(data)
 
 """
-    gettiles([src::DataSource]; lat, lon, limit = 100)
+    gettiles([src::DataSource]; lat, lon, crs = "EPSG:4326", limit = 100)
 
 Query the available data sources or a specific source `src` for tiles of
 point-cloud data. This returns an array of `PointCloudTile`s, which only
@@ -29,12 +30,13 @@ constructor to download and load the point data itself. They can also be passed
 to `minimum`/`maximum`/`extent` to get the coordinate range covered by each
 tile.
 
-The `lat` and `lon` keyword arguments specify the latitude and longitude values
-for the spatial query, specified in the [WGS
+The `lon` (or equivalently `x`) and `lat` (or `y`) keyword arguments specify
+the latitude and longitude values for the spatial query, specified in the [WGS
 84](https://en.wikipedia.org/wiki/World_Geodetic_System) coordinate reference
-system, which is also used by GPS and many online maps. The query will search
-for tiles at a point, within a box, or within a polygon, depending on whether
-one, two, or multiple coordinate values are provided.
+system (CRS), which is also used by GPS and many online maps. The query will
+search for tiles at a point, within a box, or within a polygon, depending on
+whether one, two, or multiple coordinate values are provided. Coordinates can
+also be specified in a different CRS by supplying the `crs` argument.
 
 The `limit` argument specifies the maximum number of tiles that are requested
 from each data source. The APIs that are queried may have their own limits that
@@ -45,11 +47,14 @@ by `gettiles`.
 gettiles(src; kws...) = error("No search function found for `$src`")
 
 # function should be extended for individual data sources
-gettiles(src, loc::Tuple{Real,Real}) = gettiles(src; lat = loc[1], lon = loc[2])
-gettiles(loc::Tuple{Real,Real}) = gettiles(loc...)
-function gettiles(lat::Real, lon::Real)
+function gettiles(src, loc::Tuple{Real,Real}; kws...)
+  gettiles(src; lat = loc[1], lon = loc[2], kws...)
+end
+gettiles(loc::Tuple{Real,Real}; kws...) = gettiles(; lat = loc[1], lon = loc[2], kws...)
+gettiles(lat::Real, lon::Real; kws...) = gettiles(; lat, lon, kws...)
+function gettiles(; kws...)
   tiles = PointCloudTile[]
-  append!(tiles, gettiles.(SOURCES; lat = lat, lon = lon)...)
+  append!(tiles, gettiles.(SOURCES; kws...)...)
   tiles
 end
 
