### 2025-04-30 dac6bb7
Allow downloads without SSL verification
