### 2025-05-02 90875d5
Fix & test `size(::RasterizedPointCloud)`
