### 2025-05-02 07dbbb6
Compute extrema of point clouds & rasters
