### 2025-05-02 0d55d05
Fix docstrings for `extent` argument
