### 2025-05-02 cb2cca9
Include all boundaries in rasterization extent
