### 2025-05-02 0567ccf
Fix rasterization without extent limit
