+++ test/pointcloud_test.jl
@@ -73,12 +73,16 @@ function run_pointcloud_tests(; verbose)
   r = rasterize(pts, (3, 3); extent = ((0, 0), (7, 7)))
   @test eltype(eltype(r.x)) == Float64
   @test size(r) == (3, 3)
-  count = map(xs -> length(xs), r.x)
-  @test count == [2 0 0; 0 2 0; 0 0 1]
+  @test length.(r.x) == [2 0 0; 0 2 0; 0 0 1]
   mean_height = map(r.z) do zs
     isempty(zs) ? missing : sum(zs) / length(zs)
   end
   @test eltype(mean_height) == Union{Float64,Missing}
+
+  # rasterize without extent
+  r2 = rasterize(pts, (3, 3))
+  @test extrema(r2) == ((1, 1), (5, 5))
+  @test length.(r2.x) == [2 0 0; 0 1 0; 0 0 2]
 
   # rasterize by footprint
   rf = rasterize(pts, (3, 3); extent = ((0, 0), (7, 7)), radius = 1e-3)
