### 2025-05-02 d94242c
Bump version to 0.4.2
