### 2025-05-02 efpl-columbia/PointClouds.jl#5 @mfsch
I was thinking it would be nice to support COPC and perhaps EPT as well, we just didn’t have a pressing need so far. I was unaware of that Planetary Computer dataset though, which would be really nice to have access to, especially with the ScienceBase reliability issues outlined in #4, so maybe it’s worth prioritizing COPC support.

I haven’t looked at those C++ libraries and haven’t worked with CxxWrap.jl, but that could be good option and also help with #2. I did look a bit into reimplementing the LAZ compression in Julia, but I’m not sure the specs are detailed enough, so I guess this would have to start with a 1:1 port of the laszip code (at least the arithmetic coding stuff) that can then be cleaned up to make it more idiomatic. That looked doable to me but would be a few days of work probably.
