### 2025-05-02 efpl-columbia/PointClouds.jl#4 @mfsch
I was hoping that this was perhaps a quick fix on the USGS side, but it looks like it may take a while. In the meantime I’ve added an `insecure = true` argument to the `LAS` constructor that allows skipping the SSL verification – that way we are not just silently skipping the certificate checks (6b16a99). I’ve also updated the [tutorial](https://docs.mfsch.dev/PointClouds.jl/tutorial#Finding-and-loading-LAS-data) to make use of this and mentioned it in the [docs for downloading data](https://docs.mfsch.dev/PointClouds.jl/data-sources#Downloading-point-cloud-data).
