### 2025-05-02 0eda02d
Allow skipping doctests in `just servedocs`
