### 2025-05-07 7e50d22
Add missing word in affiliation footnote
