+++ joss/references.bib
@@ -1,3 +1,10 @@
+@manual{ASPRS:2019,
+  title = {LAS Specification 1.4 -- R15},
+  year = {2019},
+  author = {{American Society for Photogrammetry and Remote Sensing}},
+  url = {https://publicdocuments.asprs.org/las-v14-r15-2019},
+}
+
 @article{Bezanson:2017,
   title = {Julia: A fresh approach to numerical computing},
   author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
@@ -25,6 +32,17 @@
   year = {2024},
   howpublished = {Presented at the AMS Annual Meeting, 28 Jan – 1 Feb, Baltimore, MD},
   publisher = {American Meteorological Society}
+}
+
+@article {Isenburg:2013,
+  author = {Isenburg, Martin},
+  title = {LASzip},
+  year = {2013},
+  journal = {Photogrammetric Engineering & Remote Sensing},
+  volume = {79},
+  number = {2},
+  pages = {209-217},
+  doi = {10.14358/PERS.79.2.209},
 }
 
 @software{PDAL:2024,
