### 2025-05-07 f615fdf
Expand on Julia ecosystem in JOSS draft
