### 2025-05-07 a8ad5d2
Add missing pages to TOC of documentation
