### 2025-05-07 efpl-columbia/PointClouds.jl#6 @mfsch
It would be great to integrate further with the JuliaGeo ecosystem where this unlocks valuable functionality. I’m a bit hesitant to give this a full-throated endorsement though as I see the integration with other packages as a bit of a balancing act: On one hand, it is easier to combine functionality from multiple packages if they form a cohesive ecosystem, and allows building up high-quality packages for shared functionality that is perhaps not at the heart of each individual project. On the other hand, I value lean dependency chains, and having parts of the functionality spread out over many packages can make it hard to advance the development as a cohesive whole, especially if the packages have different priorities/values (e.g. performance vs composability). I’ve also found that some integrations can make things harder when you’re not using the same set of packages and have to go hunting for a way to get the raw data out of some custom data structure.

To balance these trade-offs, I like having specific workflows in mind, such as your example of saving to geotiff. Package extensions also make things a bit easier, since you can keep the integrations optional. If anyone has examples of interactions with other packages that are currently cumbersome, it would be great to collect them here!
