### 2025-05-07 efpl-columbia/PointClouds.jl#6 @mfsch
Some initial comments/thoughts on those specific packages:

- GeoInterface.jl seems like a great candidate to integrate with. I haven’t used it much, but supporting it should be not be very intrusive from what I can tell. One doubt I have is that I feel that many packages accepting point collections would expect smaller/sparser data than a typical point-cloud tile? For example, just plotting the points of a 3DEP tile (usually a few millions) in Makie does not produce a very useful plot.
- I’ve looked at GeoFormatTypes.jl when implementing some of the LAS CRS functionality. The two issues I’ve come across, that made it less of a natural fit: (1) LAS can have GeoTIFF CRS data, which is not in GeoFormatTypes and would have to be added, but since it’s not just a simple string or number that is wrapped, I’m not sure it’s a good fit for the package? For example, the data could be wrapped as multiple byte vectors, but also as a dictionary or a named tuple, which is what I’m now using to represent it after parsing. (2) The EPSG codes I’m working with when handling GeoTIFF CRS data are not always representing a CRS, they can also just represent e.g. an ellipsoid or a unit, so using the EPSG type from the package everywhere is syntactically wrong. What might make sense though is accepting those types as arguments in functions that have a `crs` keyword, where `crs = EPSG(4326)` is a bit nicer than the “stringly-typed” `crs = "EPSG:4326"`.
- The `RasterizedPointCloud` is conceptually something like a `Dict{Symbol,Matrix{Vector}}` so I’m not sure how well that type specifically maps to a Raster. Currently you would use this by mapping over e.g. `rpc.z` to compute a new Matrix with a single value at each index, e.g. a height map. Those would indeed be nice to save e.g. as a geotiff, but I’d have to look into it a bit more to see what would make sense as an interface.
