### 2025-05-08 043fda9
Add use cases & related packages to docs
