### 2025-05-08 #7952 @mfsch
@evetion: Thank you for your detailed review! Your suggestions were very helpful! As you have probably seen, I’ve responded to your issues/suggestions in the PointClouds.jl repository. To address the specific points you have raised above:

- Statement of need: I think the statement of need as it appears in the paper is not the best fit for the documentation, but I have tried to better address the question “Do the authors clearly state what problems the software is designed to solve and who the target audience is?” from the review checklist. In addition to the list of features, we now also have a [section on use cases](https://docs.mfsch.dev/PointClouds.jl/#Use-cases) in the documentation.
- Contribution guidelines: We have a [short section in the README](https://github.com/efpl-columbia/PointClouds.jl#help--contributing) about contributing to the project. We have now also added a [development section](https://docs.mfsch.dev/PointClouds.jl/development/) to the documentation with some additional information.
- State of the field: We have expanded the corresponding section in the JOSS paper to include more specific references to related work within the Julia ecosystem. We have also added a [section in the documentation](https://docs.mfsch.dev/PointClouds.jl/#Related-packages) that points to related Julia packages.
- LAS/LASzip references: Great input – we have added those references.

Do you feel like these changes address your points, or is there an important aspect that is missing?
