### 2025-05-08 #7952 @evetion
> Do you feel like these changes address your points, or is there an important aspect that is missing?

Yes, thanks for your detailed reply, and the fixes for the issues I raised. I've ticked all the boxes on my end.
