### 2025-05-09 #7952 @mfsch
Thanks again, @evetion!

@kthyng: Could you confirm the next steps? From the docs, I understand that we should create a Zenodo submission of the latest version of the code, but it wasn’t clear to me if that should be the main branch (which gets published in the Julia package repository) or the JOSS branch that also contains the paper. In the latter case, we should probably rebase it on top of main so it includes all the changes made since we started the submission process? And are there any additional changes we should make to the code or the paper before archiving it? Thanks!
