### 2025-05-09 b380015
Reword attribution section in README
