### 2025-05-30 #7952 @kthyng
@mfsch Sorry for my delay. Yes, make an archive of your repo. The point is to make sure the review edits are included. It sounds like your latter option is better. I'll look at your paper next but even if we find changes to make, you won't need to update your software archive to include them since the paper will be included at its final form as a paper.
