### 2025-06-05 c61f7fe
Avoid type ambiguity for `Base.read`
