### 2025-06-24 bf28ca1
Bump version to 1.0.0
