### 2025-06-25 #7952 @mfsch
@kthyng: Sorry for the delay – I’ve now created a v1.0.0 release and archived this version on Zenodo (https://doi.org/10.5281/zenodo.15734063). The tagged version and the archive are identical and contain all the changes that were made to the code and documentation as part of the review.

The archive therefore contains everything except for the folder with the JOSS paper, which is part of a separate `joss` branch. I have rebased that branch on top of the v1.0.0 commit.

I think that should complete the above checklist. Let me know if there’s anything wrong or missing. Thank you!
