### 2025-06-27 #7952 @mfsch
@kthyng: From what I can tell, the error is related to the footnote in the affiliations – apparently the validation doesn’t like that. I’ll try to come up with a fix.
