### 2025-07-08 17bc254
Move affiliation footnote to acknowledgements
