### 2025-07-08 32a0fd6
Add ROR IDs for affiliations
