### 2025-07-08 #7952 @mfsch
@kthyng: I have reproduced the validation error locally, and moved the text from the footnote to the acknowledgements, which appears to have fixed the issue. I just pushed those changes to our `joss` branch, so the validation actions should hopefully pass now. I’ve also added ROR identifiers to the institutions.
