### 2025-07-08 #7952 @kthyng
Congratulations on your new publication @mfsch! Many thanks to reviewers @evetion and @atzberg for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
