### 2025-07-08 #7952 @mfsch
@kthyng: One little issue I just noticed: The JOSS website appears to have an issue with the ampersand in the title (see https://joss.theoj.org/papers/10.21105/joss.07952). Is there a way to get this fixed?
